Remove the extra api module/blueprint; add a standalone config module
